backend/requirements.txt
# FastAPI Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
firebase-admin==6.2.0

# Database
asyncpg==0.29.0
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Async task processing
celery==5.3.4
redis==5.0.1

# HTTP client for AI services
httpx==0.25.2
aiohttp==3.9.1

# Background processing
background-tasks==1.0.0

# CORS middleware
python-cors==1.7.0

# Logging and monitoring
structlog==23.2.0

# File upload handling
python-magic==0.4.27

# Date/time utilities
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# PostgreSQL with PostGIS support
psycopg2-binary==2.9.9

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1